--- a/requirements.txt
+++ b/requirements.txt
@@ -4,5 +4,5 @@ tokenizers>=0.13,<1
 onnxruntime>=1.14,<2 
 pyannote-audio>=3.1.1
 torch>=2.1.1 
-torchaudio>=2.1.2
+av>=11
 tqdm