This PR reverts the torchaudio code that was added in #856 and removes `torchaudio` dependency but still keeps `torch`

the reason why `torch` wasn't removed in this PR is that feature extraction still depends on it and I didn't want to include the numpy feature extraction in this PR to keep it simple and to reduce the number of conflicts to be resolved with #936 

this should be merged before #936, after both are merged, a new PR will be created to completely remove the `torch` dependency