@joiemoie 2024-08-16
Will removing torch remove the supposed FFT speedup?