@MahmoudAshraf97 2024-08-16
> Will removing torch remove the supposed FFT speedup?

Yes, but I have a new implementation using numpy in progress that is faster than the original implementation 
These are the performance figures on my machine for a 30s segment:
Current Torch on cpu vs new numpy vs old numpy
![image_20240817_022356948a5b27-82c6-46cd-a932-266a7b0610f8](https://github.com/user-attachments/assets/3c73fd34-b950-41ae-a543-53aee11d8d22)

I'm still investigating the possibility of GPU acceleration