@joiemoie 2024-08-16
Great to hear! I work with segments about 10 seconds long, so no benefit from batching. However, I am curious and possibly interested into bumping up to the latest commit because of this FFT speedup