@MahmoudAshraf97 2024-08-17
these are the latest performance figures for the new FE @joiemoie 
```
# torch on CPU
4.7 ms ± 179 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# torch on GPU
689 µs ± 9.41 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# New numpy on CPU
30.7 ms ± 1.4 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
# New CuPY on GPU
5.31 ms ± 138 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# Old numpy on CPU
61.1 ms ± 2.86 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```