@ozancaglayan 2024-08-18
Do you take into account the overhead of moving long audio tensors to GPU when doing the above measurements on GPU?