@ozancaglayan 2024-08-18
Another option could actually be to move the feature extraction layer to CTranslate2 as well, whisper.cpp has some implementations
https://github.com/ggerganov/whisper.cpp/blob/master/src/whisper-mel-cuda.cu